+++ lib/l10n/app_localizations_es.dart
@@ -1021,4 +1021,13 @@ class AppLocalizationsEs extends AppLocalizations {
 
   @override
   String get forked => 'Forked';
+
+  @override
+  String get editIssue => 'Edit Issue';
+
+  @override
+  String get titleRequired => 'Title is required';
+
+  @override
+  String get selectMilestone => 'Select milestone';
 }